move SpyderPdb to new file
